qa: increase the http.maxRequestBuffer to 100MB
Set to change the largest ref negotiation request that git will
handle during a fetch. The default value is 10MB and any fetch
requiring a larger buffer will not succeed.

This value should not normally need to be changed, but may be
helpful if you are fetching from a repository with an extremely
large number of refs.

Fixes: https://tracker.ceph.com/issues/66991
Signed-off-by: Xiubo Li <xiubli@redhat.com>