python-common/deployment: disable some unused import flake8 errors
Tell flake8 not to flag unused imports when those imports are part of an
old-style type comment. This was done rather than updating the type
annotations to the latest style because I am less familiar with this
file and there are many type comments.

Part of an effort to get ceph tox environments passing on Python 3.12.

Signed-off-by: John Mulligan <jmulligan@redhat.com>