doc/start: update os-recommendations.rst
Remove information about the operating systems that support Ceph's
official container images from the "Platforms" table in
doc/start/os-recommendations.rst and add that information to the (new)
table that shows the operating systems that support Ceph's official
container images.

Credit for this change should go to Enrico Bocchi, who noticed a
discrepancy that motivated it.

Signed-off-by: Zac Dover <zac.dover@proton.me>